refactoring - 2 new  files namespaces_store and logical_system_namespaces, and other minor changes
